requirements.txt
Appium-Python-Client
frida-tools
tidevice[openssl]
paramiko
mpv
PyQt5
PyQt5-stubs
git+https://github.com/miticollo/ios-screen-record.git
pymobiledevice3
coloredlogs
mitmproxy